+++ opencascade-hs/cpp/hs_BRepBndLib.h
@@ -9,6 +9,10 @@ extern "C" {
 
 void hs_BRepBndLib_add(TopoDS_Shape * shape, Bnd_Box * box, bool useTriangulation);
 
+void hs_BRepBndLib_addOptimal(TopoDS_Shape * shape, Bnd_Box * box, bool useTriangulation, bool useShapeTolerance);
+
+void hs_BRepBndLib_addOBB(TopoDS_Shape *shape, Bnd_OBB * obb, bool isTriangulationUsed, bool isOptimal, bool isShapeToleranceUsed);
+
 #ifdef __cplusplus
 }
 #endif