+++ opencascade-hs/src/OpenCascade/Bnd/Types.hs
@@ -1,6 +1,8 @@
 {-# LANGUAGE EmptyDataDecls #-}
 module OpenCascade.Bnd.Types
 ( Box
+, OBB
 ) where
 
 data Box
+data OBB